+++ src/App.tsx
@@ -1,39 +1,29 @@
 import "./App.css";
-import { skills } from "./data/skills";
-import { Skill } from "./types/skill";
+import SkillForm from "./components/skillformular";
+import TestTypeScript from "./TestTypeScript";
+import Diary from "./components/Diary"
+import { Routes,Route, Link } from "react-router-dom";
+import SkillsOverview from "./components/SkillsOverview";
 
 function App() {
-  const tablerows = skills.map((skill: Skill) => (
-    <tr>
-      <td>{skill.name}</td>
-      <td>{skill.level}</td>
-      <td>{convertDate(skill.created)}</td>
-    </tr>
-  ));
-
-  function convertDate(date: Date): string {
-    const day = date.getDate().toString().padStart(2, "0");
-    const month = (date.getMonth() + 1).toString().padStart(2, "0");
-    const year = date.getFullYear();
-
-    const formattedDate = day + "." + month + "." + year;
-
-    return formattedDate;
-  }
 
   return (
     <>
-      <h1>dev skill tracker</h1>
-      <table>
-        <thead>
-          <tr>
-            <th>Skill</th>
-            <th>Level</th>
-            <th>Started</th>
-          </tr>
-        </thead>
-        <tbody>{tablerows}</tbody>
-      </table>
+    <div className="App">
+      <nav className="nav">
+        <Link to="/" className="nav-item">Skills Overview</Link>
+        <Link to="/diary" className="nav-item">Diary</Link>
+      </nav>
+      <Routes>
+        <Route path="/" element={<>
+        <SkillForm />
+        <SkillsOverview />
+        <TestTypeScript />
+        </>} />
+        <Route path="/diary" element={<Diary />} />
+      </Routes>
+
+    </div>
     </>
   );
 }
